chore: move deep link router to folder
